+++ src/viur/core/bones/base.py
@@ -6,15 +6,18 @@ built, such as string, numeric, and date/time bones.
 """
 
 import copy
+import dataclasses
+import enum
 import hashlib
 import inspect
 import logging
-from dataclasses import dataclass, field
-from datetime import datetime, timedelta
-from collections.abc import Iterable
-from enum import Enum
 import typing as t
-from viur.core import db, utils, current, i18n
+from collections.abc import Iterable
+from dataclasses import dataclass, field
+from datetime import timedelta
+from enum import Enum
+
+from viur.core import current, db, i18n, utils
 from viur.core.config import conf
 
 if t.TYPE_CHECKING:
@@ -203,6 +206,55 @@ class Compute:
     raw: bool = True  # defines whether the value returned by fn is used as is, or is passed through bone.fromClient
 
 
+class CloneStrategy(enum.StrEnum):
+    """Strategy for selecting the value of a cloned skeleton"""
+
+    SET_NULL = enum.auto()
+    """Sets the cloned bone value to None."""
+
+    SET_DEFAULT = enum.auto()
+    """Sets the cloned bone value to its defaultValue."""
+
+    SET_EMPTY = enum.auto()
+    """Sets the cloned bone value to its emptyValue."""
+
+    COPY_VALUE = enum.auto()
+    """Copies the bone value from the source skeleton."""
+
+    CUSTOM = enum.auto()
+    """Uses a custom-defined logic for setting the cloned value.
+    Requires :attr:`CloneBehavior.custom_func` to be set.
+    """
+
+
+class CloneCustomFunc(t.Protocol):
+    """Type for a custom clone function assigned to :attr:`CloneBehavior.custom_func`"""
+
+    def __call__(self, skel: "SkeletonInstance", src_skel: "SkeletonInstance", bone_name: str) -> t.Any:
+        """Return the value for the cloned bone"""
+        ...
+
+
+@dataclass
+class CloneBehavior:
+    """Strategy configuration for selecting the value of a cloned skeleton"""
+
+    strategy: CloneStrategy
+    """The strategy used to select a value from a cloned skeleton"""
+
+    custom_func: CloneCustomFunc = None
+    """custom-defined logic for setting the cloned value
+    Only required when :attr:`strategy` is set to :attr:`CloneStrategy.CUSTOM`.
+    """
+
+    def __post_init__(self):
+        """Validate this configuration."""
+        if self.strategy == CloneStrategy.CUSTOM and self.custom_func is None:
+            raise ValueError("CloneStrategy is CUSTOM, but custom_func is not set")
+        elif self.strategy != CloneStrategy.CUSTOM and self.custom_func is not None:
+            raise ValueError("custom_func is set, but CloneStrategy is not CUSTOM")
+
+
 class BaseBone(object):
     """
     The BaseBone class serves as the base class for all bone types in the ViUR framework.
@@ -259,6 +311,7 @@ class BaseBone(object):
         unique: None | UniqueValue = None,
         vfunc: callable = None,  # fixme: Rename this, see below.
         visible: bool = True,
+        clone_behavior: CloneBehavior | CloneStrategy | None = None,
     ):
         """
         Initializes a new Bone.
@@ -374,6 +427,19 @@ class BaseBone(object):
             self._prevent_compute = False
 
         self.compute = compute
+
+        if clone_behavior is None:  # auto choose
+            if self.unique and self.readOnly:
+                self.clone_behavior = CloneBehavior(CloneStrategy.SET_DEFAULT)
+            else:
+                self.clone_behavior = CloneBehavior(CloneStrategy.COPY_VALUE)
+            # TODO: Any different setting for computed bones?
+        elif isinstance(clone_behavior, CloneStrategy):
+            self.clone_behavior = CloneBehavior(strategy=clone_behavior)
+        elif isinstance(clone_behavior, CloneBehavior):
+            self.clone_behavior = clone_behavior
+        else:
+            raise TypeError(f"'clone_behavior' must be an instance of Clone, but {clone_behavior=} was specified")
 
     def __set_name__(self, owner: "Skeleton", name: str) -> None:
         self.skel_cls = owner
@@ -1147,7 +1213,10 @@ class BaseBone(object):
 
         return query
 
-    def _hashValueForUniquePropertyIndex(self, value: str | int) -> list[str]:
+    def _hashValueForUniquePropertyIndex(
+        self,
+        value: str | int | float | db.Key | list[str | int | float | db.Key],
+    ) -> list[str]:
         """
         Generates a hash of the given value for creating unique property indexes.
 
@@ -1155,16 +1224,17 @@ class BaseBone(object):
         for constructing unique property indexes. Derived bone classes should overwrite this method to
         implement their own logic for hashing values.
 
-        :param value: The value to be hashed, which can be a string, integer, or a float.
+        :param value: The value(s) to be hashed.
 
         :return: A list containing a string representation of the hashed value. If the bone is multiple,
                 the list may contain more than one hashed value.
         """
-        def hashValue(value: str | int) -> str:
+
+        def hashValue(value: str | int | float | db.Key) -> str:
             h = hashlib.sha256()
             h.update(str(value).encode("UTF-8"))
             res = h.hexdigest()
-            if isinstance(value, int) or isinstance(value, float):
+            if isinstance(value, int | float):
                 return f"I-{res}"
             elif isinstance(value, str):
                 return f"S-{res}"
@@ -1181,9 +1251,9 @@ class BaseBone(object):
 
         if not value and not self.unique.lockEmpty:
             return []  # We are zero/empty string and these should not be locked
-        if not self.multiple:
+        if not self.multiple and not isinstance(value, list):
             return [hashValue(value)]
-        # We have an multiple bone here
+        # We have a multiple bone or multiple values here
         if not isinstance(value, list):
             value = [value]
         tmpList = [hashValue(x) for x in value]
@@ -1227,13 +1297,13 @@ class BaseBone(object):
         """
         pass  # We do nothing by default
 
-    def postSavedHandler(self, skel: 'viur.core.skeleton.SkeletonInstance', boneName: str, key: str):
+    def postSavedHandler(self, skel: "SkeletonInstance", boneName: str, key: db.Key | None) -> None:
         """
             Can be overridden to perform further actions after the main entity has been written.
 
             :param boneName: Name of this bone
             :param skel: The skeleton this bone belongs to
-            :param key: The (new?) Database Key we've written to
+            :param key: The (new?) Database Key we've written to. In case of a RelSkel the key is None.
         """
         pass
 
@@ -1246,6 +1316,29 @@ class BaseBone(object):
             :param key: The old Database Key of the entity we've deleted
         """
         pass
+
+    def clone_value(self, skel: "SkeletonInstance", src_skel: "SkeletonInstance", bone_name: str) -> None:
+        """Clone / Set the value for this bone depending on :attr:`clone_behavior`"""
+        match self.clone_behavior.strategy:
+            case CloneStrategy.COPY_VALUE:
+                try:
+                    skel.accessedValues[bone_name] = copy.deepcopy(src_skel.accessedValues[bone_name])
+                except KeyError:
+                    pass  # bone_name is not in accessedValues, cannot clone it
+                try:
+                    skel.renderAccessedValues[bone_name] = copy.deepcopy(src_skel.renderAccessedValues[bone_name])
+                except KeyError:
+                    pass  # bone_name is not in renderAccessedValues, cannot clone it
+            case CloneStrategy.SET_NULL:
+                skel.accessedValues[bone_name] = None
+            case CloneStrategy.SET_DEFAULT:
+                skel.accessedValues[bone_name] = self.getDefaultValue(skel)
+            case CloneStrategy.SET_EMPTY:
+                skel.accessedValues[bone_name] = self.getEmptyValue()
+            case CloneStrategy.CUSTOM:
+                skel.accessedValues[bone_name] = self.clone_behavior.custom_func(skel, src_skel, bone_name)
+            case other:
+                raise NotImplementedError(other)
 
     def refresh(self, skel: 'viur.core.skeleton.SkeletonInstance', boneName: str) -> None:
         """
@@ -1299,7 +1392,7 @@ class BaseBone(object):
         the value is valid. If the value is invalid, no modification occurs. The function supports appending values to
         bones with multiple=True and setting or appending language-specific values for bones that support languages.
         """
-        assert not (bool(self.languages) ^ bool(language)), "Language is required or not supported"
+        assert not (bool(self.languages) ^ bool(language)), f"language is required or not supported on {boneName!r}"
         assert not append or self.multiple, "Can't append - bone is not multiple"
 
         if not append and self.multiple:
@@ -1447,14 +1540,17 @@ class BaseBone(object):
         ret = {
             "descr": self.descr,
             "type": self.type,
-            "required": self.required,
+            "required": self.required and not self.readOnly,
             "params": self.params,
             "visible": self.visible,
             "readonly": self.readOnly,
             "unique": self.unique.method.value if self.unique else False,
             "languages": self.languages,
             "emptyvalue": self.getEmptyValue(),
-            "indexed": self.indexed
+            "indexed": self.indexed,
+            "clone_behavior": {
+                "strategy": self.clone_behavior.strategy,
+            },
         }
 
         # Provide a defaultvalue, if it's not a function.
@@ -1470,6 +1566,8 @@ class BaseBone(object):
             }
         else:
             ret["multiple"] = self.multiple
+
+        # Provide compute information
         if self.compute:
             ret["compute"] = {
                 "method": self.compute.interval.method.name