+++ src/viur/core/i18n.py
@@ -74,12 +74,14 @@ Of course you can create skeletons / entries in the datastore in your project
 on your own. Just use the TranslateSkel).
 """  # FIXME: grammar, rst syntax
 import datetime
-import fnmatch
 import jinja2.ext as jinja2
 import logging
+import sys
 import traceback
 import typing as t
 from pathlib import Path
+
+import jinja2.ext as jinja2
 
 from viur.core import current, db, languages, tasks
 from viur.core.config import conf
@@ -159,6 +161,8 @@ class translate:
         "force_lang",
         "public",
         "add_missing",
+        "filename",
+        "lineno",
     )
 
     def __init__(
@@ -169,6 +173,7 @@ class translate:
         force_lang: str = None,
         public: bool = False,
         add_missing: bool = False,
+        caller_is_jinja: bool = False,
     ):
         """
         :param key: The unique key defining this text fragment.
@@ -181,24 +186,42 @@ class translate:
             target language.
         :param force_lang: Use this language instead the one of the request.
         :param public: Flag for public translations, which can be obtained via /json/_translate/get_public.
+        :param caller_is_jinja: Is the call caused by our jinja method?
         """
         super().__init__()
 
         if not isinstance(key, str):
             logging.warning(f"Got non-string (type {type(key)}) as {key=}!", exc_info=True)
 
+        if force_lang is not None and force_lang not in conf.i18n.available_dialects:
+            raise ValueError(f"The language {force_lang=} is not available")
+
         key = str(key)  # ensure key is a str
         self.key = key.lower()
         self.defaultText = defaultText or key
         self.hint = hint
-
         self.translationCache = None
-        if force_lang is not None and force_lang not in conf.i18n.available_dialects:
-            raise ValueError(f"The language {force_lang=} is not available")
-
         self.force_lang = force_lang
         self.public = public
-        self.add_missing = add_missing
+        self.filename, self.lineno = None, None
+
+        if conf.i18n.add_missing_translations and self.key not in systemTranslations:
+            # This translation seems to be new and should be added
+            for frame, line in traceback.walk_stack(sys._getframe(0).f_back):
+                if self.filename is None:
+                    # Use the first frame as fallback.
+                    # In case of calling this class directly,
+                    # this is anyway the caller we're looking for.
+                    self.filename = frame.f_code.co_filename
+                    self.lineno = frame.f_lineno
+                    if not caller_is_jinja:
+                        break
+                if caller_is_jinja and not frame.f_code.co_filename.endswith(".py"):
+                    # Look for the latest html, macro (not py) where the
+                    # translate method has been used, that's our caller
+                    self.filename = frame.f_code.co_filename
+                    self.lineno = line
+                    break
 
     def __repr__(self) -> str:
         return f"<translate object for {self.key} with force_lang={self.force_lang}>"
@@ -207,51 +230,17 @@ class translate:
         if self.translationCache is None:
             global systemTranslations
 
-            from viur.core.render.html.env.viur import translate as jinja_translate
+            if conf.i18n.add_missing_translations and self.key not in systemTranslations:
+                # This translation seems to be new and should be added
 
-            if self.key not in systemTranslations:
-                # either the translate()-object has add_missing set
-                if not (add_missing := self.add_missing):
-                    # otherwise, use configuration flag
-                    add_missing = conf.i18n.add_missing_translations
-
-                    # match against fnmatch pattern, when given
-                    if isinstance(add_missing, str):
-                        add_missing = fnmatch.fnmatch(self.key, add_missing)
-                    elif isinstance(add_missing, t.Iterable):
-                        add_missing = any(fnmatch.fnmatch(self.key, pat) for pat in add_missing)
-                    else:
-                        add_missing = bool(add_missing)
-
-                if add_missing:
-                    # This translation seems to be new and should be added
-                    filename = lineno = None
-                    is_jinja = False
-                    for frame, line in traceback.walk_stack(None):
-                        if filename is None:
-                            # Use the first frame as fallback.
-                            # In case of calling this class directly,
-                            # this is anyway the caller we're looking for.
-                            filename = frame.f_code.co_filename
-                            lineno = frame.f_lineno
-                        if frame.f_code == jinja_translate.__code__:
-                            # The call was caused by our jinja method
-                            is_jinja = True
-                        if is_jinja and not frame.f_code.co_filename.endswith(".py"):
-                            # Look for the latest html, macro (not py) where the
-                            # translate method has been used, that's our caller
-                            filename = frame.f_code.co_filename
-                            lineno = line
-                            break
-
-                    add_missing_translation(
-                        key=self.key,
-                        hint=self.hint,
-                        default_text=self.defaultText,
-                        filename=filename,
-                        lineno=lineno,
-                        public=self.public,
-                    )
+                add_missing_translation(
+                    key=self.key,
+                    hint=self.hint,
+                    default_text=self.defaultText,
+                    filename=self.filename,
+                    lineno=self.lineno,
+                    public=self.public,
+                )
 
             self.translationCache = self.merge_alias(systemTranslations.get(self.key, {}))
 
@@ -360,7 +349,7 @@ class TranslationExtension(jinja2.Extension):
         tr_key = args[0].lower()
         public = kwargs.pop("_public_", False) or False
 
-        if tr_key not in systemTranslations:
+        if conf.i18n.add_missing_translations and tr_key not in systemTranslations:
             add_missing_translation(
                 key=tr_key,
                 hint=args[1],